refactor: refactor to FSD structure
