+++ src/providers/ApolloProvider.tsx
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import { ApolloProvider } from '@apollo/client/react';
-import { client } from '@/api/apollo-client';
+import { client } from '@/shared/api/apollo-client';
 
 export const ApolloClientProvider = ({ children }: { children: React.ReactNode }) => {
   return <ApolloProvider client={client}>{children}</ApolloProvider>;
