+++ src/widgets/app-layout/StyledLayout.tsx
